Fix integer conversion warning

+++ src/http.c
@@ -9,7 +9,7 @@ static const char *const kHTTPVersions[] = {
 };
 
 HTTPVersion GetHTTPVersion(const char *http_version_string) {
-    for (size_t i = 0; i < sizeof kHTTPVersions / sizeof(const char *); ++i) {
+    for (int i = 0; i < sizeof kHTTPVersions / sizeof(const char *); ++i) {
         if (strcmp(http_version_string, kHTTPVersions[i]) == 0) {
             return i;
         }
@@ -31,7 +31,7 @@ static const char *const kRequestMethods[] = {
 };
 
 RequestMethod GetRequestMethod(const char *request_method_string) {
-    for (size_t i = 0; i < sizeof kRequestMethods / sizeof(const char *); ++i) {
+    for (int i = 0; i < sizeof kRequestMethods / sizeof(const char *); ++i) {
         if (strcmp(request_method_string, kRequestMethods[i]) == 0) {
             return i;
         }
